final/book/requirements/django.txt
Django==1.10
django-braces==1.11.0
django-formset-js==0.5.0
django-jquery-js==3.1.1
django-model-utils==3.0.0
django-star-ratings==0.5.5
djangorestframework==3.6.3
